requirements.txt
aenum==2.2.4
attrs==20.2.0
beautifulsoup4==4.9.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fbchat==1.9.7
idna==2.10
iniconfig==1.0.1
packaging==20.4
paho-mqtt==1.5.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
